Extract ChannelMessageList and remove LoggedInContext
Unify the duplicated message bubble rendering between ChannelScreen and
GuestChannelView into a shared ChannelMessageList component. Remove the
LoggedInContext indirection in Router — auth guard components now call
useAuth() directly since AuthProvider wraps the full tree.
