+++ packages/web/src/App.tsx
@@ -5,7 +5,7 @@ import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import Router from "./Router";
 import { MaintenanceMode } from "./components/MaintenanceMode";
 import { ErrorBoundary } from "./components/ErrorBoundary";
-import { AuthProvider, useAuth } from "./auth";
+import { AuthProvider } from "./auth";
 import { Toaster } from "./components/ui/sonner";
 import { isNativePlatform } from "./utils/platform";
 import { initializeNativeSocialLogin } from "./auth/nativeGoogleAuth";
@@ -33,14 +33,12 @@ const NotificationPermissionPrompter: React.FC = () => {
 };
 
 function AppContent() {
-  const { loggedIn } = useAuth();
-
   return (
     <>
       <Suspense fallback={null}>
         <NotificationPermissionPrompter />
       </Suspense>
-      <Router loggedIn={loggedIn} queryClient={queryClient} />
+      <Router queryClient={queryClient} />
     </>
   );
 }
