+++ packages/web/src/screens/GameDetail/OrdersScreen.tsx
@@ -14,7 +14,6 @@ import {
   UserX,
   Handshake,
   LucideIcon,
-  MessageCircle,
   ChevronRight,
 } from "lucide-react";
 import { toast } from "sonner";
@@ -41,8 +40,8 @@ import {
 } from "@/components/ui/item";
 import { Notice } from "@/components/Notice";
 import { NationFlag, findNationFlagUrl, findNationColor } from "@/components/NationFlag";
-import { Message, MessageContent, MessageTimestamp } from "@/components/ui/message";
-import { getChannelFlagUrls, brightnessByColor, toHex6 } from "./channelUtils";
+import { getChannelFlagUrls } from "./channelUtils";
+import { buildMessageDisplayItems, ChannelMessageList } from "./ChannelMessageList";
 import { ChannelAvatar } from "./ChannelAvatar";
 import { GameDropdownMenu } from "@/components/GameDropdownMenu";
 import { GameDetailAppBar } from "./AppBar";
@@ -55,7 +54,6 @@ import {
   PhaseState,
   Province,
   Channel,
-  ChannelMessage as ChannelMessageType,
   useGameOrdersDeleteDestroy,
   useGameOrdersListSuspense,
   useGamePhaseRetrieveSuspense,
@@ -77,38 +75,6 @@ import {
 } from "@/api/generated/endpoints";
 import { cn } from "../../lib/utils";
 
-const BUBBLE_ALPHA_HEX = "26";
-
-type MessageDisplayItem = {
-  id: number;
-  body: string;
-  sender: { nationName: string; nationColor: string };
-  isCurrentUser: boolean;
-  showAvatar: boolean;
-  formattedTime: string;
-};
-
-const formatMessageTime = (createdAt: string): string => {
-  const date = new Date(createdAt);
-  const today = new Date();
-  const isToday =
-    date.getFullYear() === today.getFullYear() &&
-    date.getMonth() === today.getMonth() &&
-    date.getDate() === today.getDate();
-  const time = date.toLocaleTimeString([], { hour: "2-digit", minute: "2-digit" });
-  if (isToday) return time;
-  return `${date.toLocaleDateString([], { month: "short", day: "numeric" })} ${time}`;
-};
-
-const buildMessageDisplayItems = (messages: readonly ChannelMessageType[]): MessageDisplayItem[] =>
-  messages.map((msg, index) => ({
-    id: msg.id,
-    body: msg.body,
-    sender: { nationName: msg.sender.nation.name, nationColor: msg.sender.nation.color },
-    isCurrentUser: msg.sender.isCurrentUser,
-    showAvatar: index === 0 || messages[index - 1].sender.nation.name !== msg.sender.nation.name,
-    formattedTime: formatMessageTime(msg.createdAt),
-  }));
 
 interface GuestChannelViewProps {
   channel: Channel;
@@ -135,49 +101,10 @@ const GuestChannelView: React.FC<GuestChannelViewProps> = ({ channel, game, vari
         <Panel>
           <Panel.Content>
             <div className="h-full flex flex-col">
-              {messageItems.length === 0 ? (
-                <Notice icon={MessageCircle} title="No messages yet" className="h-full" />
-              ) : (
-                <div className="flex-1 min-h-0 overflow-y-auto flex flex-col gap-2 p-2">
-                  {messageItems.map(item => (
-                    <Message key={item.id} className={item.isCurrentUser ? "flex-row-reverse" : undefined}>
-                      {item.showAvatar ? (
-                        <div className="w-8 flex-shrink-0 flex justify-center">
-                          <NationFlag
-                            flagUrl={findNationFlagUrl(variant.nations, item.sender.nationName)}
-                            alt={item.sender.nationName}
-                            size="lg"
-                            color={item.sender.nationColor}
-                          />
-                        </div>
-                      ) : (
-                        <div className="w-8 flex-shrink-0" />
-                      )}
-                      <MessageContent
-                        className={`py-1.5 px-2 ${item.isCurrentUser ? "rounded-tr-none" : "rounded-tl-none"}`}
-                        style={{
-                          backgroundColor: toHex6(item.sender.nationColor) + BUBBLE_ALPHA_HEX,
-                          border: brightnessByColor(item.sender.nationColor) > 128
-                            ? `1px solid ${item.sender.nationColor}`
-                            : undefined,
-                        }}
-                      >
-                        {item.body}
-                        {item.showAvatar ? (
-                          <div className="flex items-center justify-between gap-2 mt-0.5">
-                            <span className="text-xs font-medium" style={{ color: item.sender.nationColor }}>
-                              {item.sender.nationName}
-                            </span>
-                            <span className="text-xs text-muted-foreground">{item.formattedTime}</span>
-                          </div>
-                        ) : (
-                          <MessageTimestamp className="mt-0.5">{item.formattedTime}</MessageTimestamp>
-                        )}
-                      </MessageContent>
-                    </Message>
-                  ))}
-                </div>
-              )}
+              <ChannelMessageList
+                messageItems={messageItems}
+                variantNations={variant.nations}
+              />
             </div>
           </Panel.Content>
         </Panel>
